Move utility functions and type definitions to separate modules
