+++ src/crewai/rag/chromadb/client.py
@@ -1,205 +1,32 @@
 """ChromaDB client implementation."""
 
-import hashlib
-from collections.abc import Mapping
-from typing import Any, TypeGuard
+from typing import Any
 
-from chromadb.api import AsyncClientAPI, ClientAPI
-from chromadb.api.configuration import CollectionConfigurationInterface
 from chromadb.api.types import (
-    CollectionMetadata,
-    DataLoader,
     Embeddable,
     EmbeddingFunction as ChromaEmbeddingFunction,
-    Include,
-    IncludeEnum,
-    Loadable,
     QueryResult,
-    Where,
-    WhereDocument,
 )
 from typing_extensions import Unpack
 
 from crewai.rag.chromadb.types import (
     ChromaDBClientType,
-    PreparedDocuments,
-    ExtractedSearchParams,
+    ChromaDBCollectionCreateParams,
+    ChromaDBCollectionSearchParams,
+)
+from crewai.rag.chromadb.utils import (
+    _convert_chromadb_results_to_search_results,
+    _extract_search_params,
+    _is_async_client,
+    _is_sync_client,
+    _prepare_documents_for_chromadb,
 )
 from crewai.rag.core.base_client import (
     BaseClient,
     BaseCollectionParams,
     BaseCollectionAddParams,
-    BaseCollectionSearchParams,
 )
-from crewai.rag.types import (
-    BaseRecord,
-    SearchResult,
-)
-
-
-def _is_sync_client(client: ChromaDBClientType) -> TypeGuard[ClientAPI]:
-    """Type guard to check if the client is a synchronous ClientAPI.
-
-    Args:
-        client: The client to check.
-
-    Returns:
-        True if the client is a ClientAPI, False otherwise.
-    """
-    return isinstance(client, ClientAPI)
-
-
-def _is_async_client(client: ChromaDBClientType) -> TypeGuard[AsyncClientAPI]:
-    """Type guard to check if the client is an asynchronous AsyncClientAPI.
-
-    Args:
-        client: The client to check.
-
-    Returns:
-        True if the client is an AsyncClientAPI, False otherwise.
-    """
-    return isinstance(client, AsyncClientAPI)
-
-
-class ChromaDBCollectionCreateParams(BaseCollectionParams, total=False):
-    """Parameters for creating a ChromaDB collection.
-
-    This class extends BaseCollectionParams to include any additional
-    parameters specific to ChromaDB collection creation.
-    """
-
-    configuration: CollectionConfigurationInterface
-    metadata: CollectionMetadata
-    embedding_function: ChromaEmbeddingFunction[Embeddable]
-    data_loader: DataLoader[Loadable]
-    get_or_create: bool
-
-
-class ChromaDBCollectionSearchParams(BaseCollectionSearchParams, total=False):
-    """Parameters for searching a ChromaDB collection.
-
-    This class extends BaseCollectionSearchParams to include ChromaDB-specific
-    search parameters like where clauses and include options.
-    """
-
-    where: Where
-    where_document: WhereDocument
-    include: Include
-
-
-def _prepare_documents_for_chromadb(
-    documents: list[BaseRecord],
-) -> PreparedDocuments:
-    """Prepare documents for ChromaDB by extracting IDs, texts, and metadata.
-
-    Args:
-        documents: List of BaseRecord documents to prepare.
-
-    Returns:
-        PreparedDocuments with ids, texts, and metadatas ready for ChromaDB.
-    """
-    ids: list[str] = []
-    texts: list[str] = []
-    metadatas: list[Mapping[str, str | int | float | bool]] = []
-
-    for doc in documents:
-        if "doc_id" in doc:
-            ids.append(doc["doc_id"])
-        else:
-            content_hash = hashlib.sha256(doc["content"].encode()).hexdigest()[:16]
-            ids.append(content_hash)
-
-        texts.append(doc["content"])
-        metadata = doc.get("metadata")
-        if metadata:
-            if isinstance(metadata, list):
-                metadatas.append(metadata[0] if metadata else {})
-            else:
-                metadatas.append(metadata)
-        else:
-            metadatas.append({})
-
-    return PreparedDocuments(ids, texts, metadatas)
-
-
-def _extract_search_params(
-    kwargs: ChromaDBCollectionSearchParams,
-) -> ExtractedSearchParams:
-    """Extract search parameters from kwargs.
-
-    Args:
-        kwargs: Keyword arguments containing search parameters.
-
-    Returns:
-        ExtractedSearchParams with all extracted parameters.
-    """
-    return ExtractedSearchParams(
-        collection_name=kwargs["collection_name"],
-        query=kwargs["query"],
-        limit=kwargs.get("limit", 10),
-        metadata_filter=kwargs.get("metadata_filter"),
-        score_threshold=kwargs.get("score_threshold"),
-        where=kwargs.get("where"),
-        where_document=kwargs.get("where_document"),
-        include=kwargs.get(
-            "include",
-            [IncludeEnum.metadatas, IncludeEnum.documents, IncludeEnum.distances],
-        ),
-    )
-
-
-def _convert_chromadb_results_to_search_results(
-    results: QueryResult,
-    include: Include,
-    score_threshold: float | None = None,
-) -> list[SearchResult]:
-    """Convert ChromaDB query results to SearchResult format.
-
-    Args:
-        results: ChromaDB query results.
-        include: List of fields that were included in the query.
-        score_threshold: Optional minimum similarity score (0-1) for results.
-
-    Returns:
-        List of SearchResult dicts containing id, content, metadata, and score.
-    """
-    search_results: list[SearchResult] = []
-
-    include_strings = [
-        item.value if isinstance(item, IncludeEnum) else item for item in include
-    ]
-
-    ids = results["ids"][0] if results.get("ids") else []
-
-    documents_list = results.get("documents")
-    documents = (
-        documents_list[0] if documents_list and "documents" in include_strings else []
-    )
-
-    metadatas_list = results.get("metadatas")
-    metadatas = (
-        metadatas_list[0] if metadatas_list and "metadatas" in include_strings else []
-    )
-
-    distances_list = results.get("distances")
-    distances = (
-        distances_list[0] if distances_list and "distances" in include_strings else []
-    )
-
-    for i, doc_id in enumerate(ids):
-        score = 1.0 - (distances[i] / 2.0) if distances and i < len(distances) else 0.0
-        if score_threshold and score < score_threshold:
-            continue
-
-        result: SearchResult = {
-            "id": doc_id,
-            "content": documents[i] if documents and i < len(documents) else "",
-            "metadata": dict(metadatas[i]) if metadatas and i < len(metadatas) else {},
-            "score": score,
-        }
-        search_results.append(result)
-
-    return search_results
+from crewai.rag.types import SearchResult
 
 
 class ChromaDBClient(BaseClient):
